Moved all sidebar code to SideBar.js
